+++ src/relax/transform/fma_rewrite.cc
@@ -18,7 +18,7 @@
  */
 /*!
  * \file src/relax/transform/fma_rewrite.cc
- * \brief Perform fused multiply add rewriting in dataflow blocks.
+ * \brief Perform fused multiply-add rewriting.
  */
 #include <tvm/relax/expr_functor.h>
 #include <tvm/relax/transform.h>
@@ -26,22 +26,23 @@
 namespace tvm {
 namespace relax {
 
-// ==================
-// EwiseFMARewriter
-// Example:
-// x0 = mul(a, b)
-// z0 = add(x0, c)
-// -->
-// z0 = ewise_fma(a, b, c)
-
-// Example 2:
-// Question: do we want to support this?
-// x0 = mul(a, add(k, b))
-// z0 = add(x0, c)
-// -->
-// lv0 = add(k, b)
-// z0 = ewise_fma(a, lv0, c)
-
+/*! \brief Rewrites the relax.add call to a relax.ewise_fma call when detecting the multiply-add
+ * pattern.
+ *
+ * Example:
+ * x0 = mul(a, b)
+ * z0 = add(x0, c)
+ * -->
+ * z0 = ewise_fma(a, b, c)
+ *
+ * Example 2:
+ * Question: do we want to support this?
+ * x0 = mul(a, add(k, b))
+ * z0 = add(x0, c)
+ * -->
+ * lv0 = add(k, b)
+ * z0 = ewise_fma(a, lv0, c)
+ */
 class EwiseFMARewriter : public ExprMutator {
   Expr VisitExpr_(const CallNode* call) override {
     Expr expr = VisitExprPostOrder_(call);
@@ -66,21 +67,100 @@ class EwiseFMARewriter : public ExprMutator {
   }
 };
 
-BindingBlock FMARewrite(const BindingBlock& block) {
+BindingBlock RewriteFMA(const BindingBlock& block) {
   return EwiseFMARewriter().VisitBindingBlock(block);
 }
 
+/*! \brief Performs multiply add fusion. The difference of EwiseFMARewriter and this
+ * EwiseFuseFMAMutator class is that this mutator generates a sub function(subgraph) whose body is a
+ * CallNode that calls to the relax.ewise_fma op, and rewrites the relax.add call in the main
+ * function to calling to the subgraph.
+ *
+ * Example:
+ * Before-transformation IRModule:
+ * def main():
+ *   x0 = mul(a, b)
+ *   z0 = add(x0, c)
+ * -->
+ * After-transformation IRModule:
+ * def ewise_fused(x, y, z):
+ *  return relax.ewise_fma(x, y, z)
+ *
+ * def main():
+ *  z0 = ewise_fused(a, b, c)
+ */
+class EwiseFuseFMAMutator : public ExprMutator {
+ public:
+  explicit EwiseFuseFMAMutator(IRModule mod) { mod_ = mod; }
+
+  IRModule Transform() {
+    for (auto& p : mod_->functions) {
+      Expr func = p.second;
+      if (func->IsInstance<FunctionNode>()) {
+        func = this->VisitExpr(func);
+      }
+      builder_->AddFunction(Downcast<BaseFunc>(func), p.first->name_hint);
+    }
+    return builder_->GetContextIRModule();
+  }
+
+  Expr VisitExpr_(const CallNode* call) override {
+    Expr expr = VisitExprPostOrder_(call);
+    call = expr.as<CallNode>();
+
+    static const Op& add_op = Op::Get("relax.add");
+    static const Op& multiply_op = Op::Get("relax.multiply");
+    static const Op& ewise_fma_op = Op::Get("relax.ewise_fma");
+
+    if (call->op == add_op) {
+      Optional<Expr> value = LookupBinding(Downcast<Var>(call->args[0]));
+      const CallNode* mul = value.as<CallNode>();
+      if (mul && mul->op == multiply_op) {
+        // construct a subgraph
+        Var x = Var("x", Downcast<Expr>(mul->args[0]->shape_), mul->args[0]->checked_type_);
+        Var y = Var("y", Downcast<Expr>(mul->args[1]->shape_), mul->args[1]->checked_type_);
+        Var z = Var("z", Downcast<Expr>(call->args[1]->shape_), call->args[1]->checked_type_);
+        Expr body = Call(ewise_fma_op, {x, y, z});
+
+        // TODO(@yuchen): avoid creating the unnecessary global_var after #136 is merged
+        GlobalVar global_var = GlobalVar("ewise_fma_fused");
+        Expr ewise_fma_fused = Function(global_var, {x, y, z}, body, call->args[1]->checked_type_);
+        Expr normalized = builder_->Normalize(ewise_fma_fused);
+        GlobalVar global_var1 =
+            builder_->AddFunction(Downcast<BaseFunc>(normalized), "ewise_fma_fused");
+
+        // construct a call to the subgraph
+        Call fma_call = Call(global_var1, {mul->args[0], mul->args[1], call->args[1]}, {}, {});
+        return std::move(fma_call);
+      }
+    }
+
+    return GetRef<Call>(call);
+  }
+
+ private:
+  IRModule mod_;
+};
+
 namespace transform {
 
-Pass FMARewrite() {
+Pass RewriteFMA() {
   runtime::TypedPackedFunc<DataflowBlock(DataflowBlock, IRModule, PassContext)> pass_func =
       [=](DataflowBlock block, IRModule m, PassContext pc) {
-        return Downcast<DataflowBlock>(FMARewrite(block));
+        return Downcast<DataflowBlock>(RewriteFMA(block));
       };
-  return CreateDataflowBlockPass(pass_func, 2, "FMARewrite", {});
+  return CreateDataflowBlockPass(pass_func, 2, "RewriteFMA", {});
 }
 
-TVM_REGISTER_GLOBAL("relax.transform.FMARewrite").set_body_typed(FMARewrite);
+TVM_REGISTER_GLOBAL("relax.transform.RewriteFMA").set_body_typed(RewriteFMA);
+
+Pass FuseFMA() {
+  runtime::TypedPackedFunc<IRModule(IRModule, PassContext)> pass_func =
+      [=](IRModule mod, PassContext pc) { return EwiseFuseFMAMutator(mod).Transform(); };
+  return CreateModulePass(pass_func, 2, "FuseFMA", {});
+}
+
+TVM_REGISTER_GLOBAL("relax.transform.FuseFMA").set_body_typed(FuseFMA);
 
 }  // namespace transform
