+++ tests/python/relax/test_transform.py
@@ -53,7 +53,7 @@ def test_fma_rewrite():
             gv2 = relax.add(gv1, y)
             return (gv0, gv1, gv2)
 
-    After = relax.transform.FMARewrite()(Before)
+    After = relax.transform.RewriteFMA()(Before)
 
     assert_structural_equal(After, Expected)
 
@@ -69,7 +69,7 @@ def test_fma_fuse():
                 relax.output(gv0)
             return gv0
 
-    After = relax.transform.FMAFuse()(Before)
+    After = relax.transform.FuseFMA()(Before)
 
     # TODO(@yuchen): add assert_structural_equal after normalization in parser
     assert len(After.get_global_vars()) == 2
